made it so customers only have 2 states
moving and frozen other states are to be implemented in the future.
