+++ GameDesign/Assets/Scripts/Managers/PointsManager.cs
@@ -59,7 +59,6 @@ public class PointsManager : MonoBehaviour
     public void TriggerEvent_IncrementPoints(int points)
     {
         instance.currentPoints += points;
-        Debug.Log($"Increasing points {points}, now is {instance.currentPoints}");
 
         if (points < 0) instance.lostPoints += points;
         if (points > 0) instance.earnedPoints += points;
