+++ GameDesign/Assets/Scripts/States/Frozen.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4d9b8b2a904583e4fb5921426afb01dc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
