+++ GameDesign/Assets/Scripts/States/MovingState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6f6781be029356045b00553235556a76
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
